fixing missing include dependency type_traits and pushing working executable. back to same issue with starting new game. going over all pure virtual function and making sure destructor are properly called and data cleared from memory. have to run a double check about my inheritence setup and confirm what need to be virtual and what need to be pure virutal

+++ SDL2-Minesweeper/headers/composite/ImGuiComplexComponent.h
@@ -17,8 +17,8 @@ namespace Toolset {
 		~ImGuiComplexComponent() override;
 		ImGuiComplexComponent& operator=(const ImGuiComplexComponent&) = delete;
 		ImGuiComplexComponent& operator=(ImGuiComplexComponent&&) = delete;
-		float getComponentWidth() override;
-		float getComponentHeight() override;
+		int getComponentWidth() override;
+		int getComponentHeight() override;
 		void add(ImGuiComponent*);
 		void remove(ImGuiComponent*);
 		void refresh() override;
