+++ SDL2-Minesweeper/headers/composite/Rect.h
@@ -4,17 +4,17 @@
 
 namespace Toolset {
     struct Rect {
-        float x = 0, y = 0, w = 0, h = 0;
+        int x = 0, y = 0, w = 0, h = 0;
         Rect(const Rect&) = delete;
         Rect(Rect&&) = delete;
         Rect() = delete;
-        Rect(const float& x, const float& y, const float& w, const float& h);
+        Rect(const int& x, const int& y, const int& w, const int& h);
         ~Rect() = default;
         Rect& operator=(const Rect&) = delete;
         Rect& operator=(Rect&&) = delete;
     };
 
-    inline Rect::Rect(const float& x, const float& y, const float& w, const float& h) : x(x), y(y), w(w), h(h)
+    inline Rect::Rect(const int& x, const int& y, const int& w, const int& h) : x(x), y(y), w(w), h(h)
     {
     }
 }
